requirements.txt
cdktf>=0.19.0
cdktf-cdktf-provider-aws>=18.0.0
PyYAML>=5.4.1
pytest>=7.0.0
constructs>=10.0.0
typeguard>=2.13.3
